Added Error Handling and Robustness. - should be able to see where error is in config. - invalid commands don't kill gestures. - evemu_do robustness increased by increasing time between consecutive   events. This prevents key latching (key repeats until interaction). - slimmed down getConfig.py to only include stuff I am currently using. - Added logic in install.sh and unistall.sh to ask before removing   configuration and input group removal. - Added another python dependancy (simplejson) in install.sh.

+++ install.sh
@@ -25,11 +25,24 @@ esac
 ## cat, stdbuf are builtin, may need to install daemonize by hand
 cat pkg_requirements | xargs -n 1 sudo ${pkgM} ${install} ${auto} 
 ## subprocess, shlex, threading, queue, time, os, sys, math  are builtin
-sudo pip3 install pathlib
+## installs on global python
+sudo pip3 install pathlib simplejson
 
-# place files in corrosponding locations
+# place files in corresponding locations
 sudo cp gestures evemu_do getConfig.py ${install_location}
-cp gestures.conf ${config_location}
+if test -f "${config-location}/gestures.conf"; then
+    echo "$FILE exist"
+    while true; do
+        echo "config file exists. override (saves backup as gestures.conf.bak)?" 
+        read -p "(y/n):" choice
+        case "$choice" in 
+          y|Y ) cp ${config_location}/gestures.conf ${config_location}/gestures.conf.bak; cp gestures.conf ${config_location}; break;;
+          n|N ) break;;
+          * ) echo "Please answer y or n.";;
+        esac
+    done
+
+fi
 cp gestures.desktop ${autostart_location}
 
 # add user to input group
